git  restructure code
