+++ servers/mcpgw/core/__init__.py
@@ -1,0 +1,9 @@
+"""
+Core functionality modules for MCPGW server.
+
+This package contains core business logic including:
+- Authentication and authorization
+- Scopes management  
+- Registry API client
+"""
+
