+++ servers/mcpgw/pyproject.toml
@@ -7,9 +7,11 @@ requires-python = ">=3.12,<3.13"
 dependencies = [
     "fastmcp>=2.0.0",
     "pydantic>=2.11.3",
+    "pydantic-settings>=2.0.0",
     "httpx>=0.27.0",
     "python-dotenv>=1.0.0",
     "websockets>=15.0.1",
+    "pyyaml>=6.0.2",
 ]
 
 [project.optional-dependencies]
@@ -21,3 +23,11 @@ embedded-search = [
     "sentence-transformers>=2.2.2",
     "scikit-learn>=1.3.0",
 ]
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+packages = ["core", "search", "tools"]
+py-modules = ["server", "config"]
